+++ include/atheme/pbkdf2.h
@@ -4,8 +4,8 @@
  *
  * Copyright (C) 2017-2018 Atheme Development Group (https://atheme.github.io/)
  *
- * Data structures and macros for the PBKDF2 crypto modules and the SASL
- * SCRAM-SHA module.
+ * Data structures and macros for the PBKDF2 crypto modules and the SASL SCRAM
+ * module.
  */
 
 #ifndef ATHEME_INC_PBKDF2_H
@@ -65,8 +65,8 @@ struct pbkdf2v2_dbentry
 {
 	unsigned char           cdg[DIGEST_MDLEN_MAX];                        // PBKDF2 Digest (Computed)
 	unsigned char           sdg[DIGEST_MDLEN_MAX];                        // PBKDF2 Digest (Stored)
-	unsigned char           ssk[DIGEST_MDLEN_MAX];                        // SCRAM-SHA ServerKey (Stored)
-	unsigned char           shk[DIGEST_MDLEN_MAX];                        // SCRAM-SHA StoredKey (Stored)
+	unsigned char           ssk[DIGEST_MDLEN_MAX];                        // SCRAM ServerKey (Stored)
+	unsigned char           shk[DIGEST_MDLEN_MAX];                        // SCRAM StoredKey (Stored)
 	unsigned char           salt[PBKDF2_SALTLEN_MAX];                     // PBKDF2 Salt
 	char                    salt64[BASE64_SIZE_STR(PBKDF2_SALTLEN_MAX)];  // PBKDF2 Salt (Base64-encoded)
 	size_t                  dl;                                           // Hash/HMAC/PBKDF2 Digest Length
@@ -74,7 +74,7 @@ struct pbkdf2v2_dbentry
 	unsigned int            a;                                            // PBKDF2v2 PRF ID (one of the above)
 	unsigned int            c;                                            // PBKDF2 Iteration Count
 	enum digest_algorithm   md;                                           // Atheme Digest Interface Algorithm ID
-	bool                    scram;                                        // Whether to use HMAC-SHA or SCRAM-SHA
+	bool                    scram;                                        // Whether to use SCRAM-type credentials
 };
 
 struct pbkdf2v2_scram_config